fix(core): migrate WASM technicals backend to rust_decimal
Refactors technicals-wasm lib.rs and dependencies to use
rust_decimal::Decimal instead of f64. Updates WASM boundaries
to pass string arrays for precise financial values as mandated
by BUG-0182. Updates TypeScript frontend integration correctly.

+++ technicals-wasm/src/indicator_settings.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 // Additional indicator settings structures for missing indicators
 
 #[derive(Serialize, Deserialize, Clone, Default)] 
@@ -27,7 +28,7 @@ pub struct IchimokuSettings {
 }
 
 #[derive(Serialize, Deserialize, Clone, Default)]
-pub struct KeltnerSettings { pub length: usize, pub multiplier: f64, pub atr_length: usize }
+pub struct KeltnerSettings { pub length: usize, pub multiplier: Decimal, pub atr_length: usize }
 
 #[derive(Serialize, Deserialize, Clone, Default)]
 pub struct DonchianSettings { pub length: usize }
